KFSPTS-26270 Start of PR feedback changes
