+++ src/main/java/edu/cornell/kfs/module/purap/jaggaer/common/xml/AuthenticationBase.java
@@ -1,0 +1,37 @@
+package edu.cornell.kfs.module.purap.jaggaer.common.xml;
+
+import jakarta.xml.bind.annotation.XmlAccessType;
+import jakarta.xml.bind.annotation.XmlAccessorType;
+import jakarta.xml.bind.annotation.XmlElement;
+import jakarta.xml.bind.annotation.XmlType;
+
+@XmlAccessorType(XmlAccessType.FIELD)
+@XmlType(name = "", propOrder = {
+    "identity",
+    "sharedSecret"
+})
+public abstract class AuthenticationBase {
+
+    @XmlElement(name = "Identity", required = true)
+    private String identity;
+
+    @XmlElement(name = "SharedSecret", required = true)
+    private String sharedSecret;
+
+    public String getIdentity() {
+        return identity;
+    }
+
+    public void setIdentity(String identity) {
+        this.identity = identity;
+    }
+
+    public String getSharedSecret() {
+        return sharedSecret;
+    }
+
+    public void setSharedSecret(String sharedSecret) {
+        this.sharedSecret = sharedSecret;
+    }
+
+}
