+++ src/main/java/edu/cornell/kfs/module/purap/jaggaer/supplier/xml/Authentication.java
@@ -1,35 +1,12 @@
 package edu.cornell.kfs.module.purap.jaggaer.supplier.xml;
 
+import edu.cornell.kfs.module.purap.jaggaer.common.xml.AuthenticationBase;
 import jakarta.xml.bind.annotation.XmlAccessType;
 import jakarta.xml.bind.annotation.XmlAccessorType;
-import jakarta.xml.bind.annotation.XmlElement;
 import jakarta.xml.bind.annotation.XmlRootElement;
-import jakarta.xml.bind.annotation.XmlType;
 
 @XmlAccessorType(XmlAccessType.FIELD)
-@XmlType(name = "", propOrder = { "identity", "sharedSecret" })
 @XmlRootElement(name = "Authentication")
-public class Authentication {
-
-    @XmlElement(name = "Identity", required = true)
-    private String identity;
-    @XmlElement(name = "SharedSecret", required = true)
-    private String sharedSecret;
-
-    public String getIdentity() {
-        return identity;
-    }
-
-    public void setIdentity(String identity) {
-        this.identity = identity;
-    }
-
-    public String getSharedSecret() {
-        return sharedSecret;
-    }
-
-    public void setSharedSecret(String sharedSecret) {
-        this.sharedSecret = sharedSecret;
-    }
+public class Authentication extends AuthenticationBase {
 
 }
